Add new library project PatientsAPI_SharedModels + move shared models from PatientsAPI to the new project + update PatientsAPI_Test and FrontendMVC to reference new project + remove duplicated models from FrontendMVC project
