+++ FrontendMVC/Views/Patients/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model FrontendMVC.Models.PatientModelAdd
+﻿@model PatientsAPI_SharedModels.PatientModelAdd
 
 <h2>Ajouter un Patient</h2>
