+++ FrontendMVC/Views/Patients/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@using FrontendMVC.Models
+﻿@using PatientsAPI_SharedModels
 @model PatientDetailsViewModel
 
 <h2>Détail du patient</h2>
