+++ FrontendMVC/Views/Patients/Update.cshtml
@@ -1,4 +1,4 @@
-﻿@model FrontendMVC.Models.PatientModelUpdate
+﻿@model PatientsAPI_SharedModels.PatientModelUpdate
 
 <h2>Modifier un Patient</h2>
