+++ PatientsAPI_SharedModels/PatientModelAdd.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace PatientsAPI.Models
+namespace PatientsAPI_SharedModels
 {
-    public class PatientModelUpdate
+    public class PatientModelAdd
     {
         [Required(ErrorMessage = "La saisie du nom du patient est obligatoire")]
         public string LastName { get; set; } = string.Empty;
